+++ packages/shared/src/shared/Utils.ts
@@ -3,6 +3,8 @@
  * See LICENSE.md in the project root for license terms and full copyright notice.
  *--------------------------------------------------------------------------------------------*/
 
+import { assert } from "@itwin/core-bentley";
+
 import type { EC } from "./Metadata.js";
 
 /**
@@ -13,11 +15,25 @@ import type { EC } from "./Metadata.js";
  * @public
  */
 export function parseFullClassName(fullClassName: string) {
-  const [schemaName, className, ...rest] = fullClassName.split(/[\.:]/);
-  if (!schemaName || !className || rest.length > 0) {
-    throw new Error(`Invalid full class name: ${fullClassName}`);
+  let idx = fullClassName.indexOf(".");
+  let includesDot = true;
+  if (idx <= 0) {
+    includesDot = false;
+    idx = fullClassName.indexOf(":");
+    if (idx <= 0) {
+      throw new Error(`Invalid full class name: ${fullClassName}`);
+    }
   }
-  return { schemaName, className };
+  assert(() => {
+    return (
+      idx < fullClassName.length - 1 &&
+      (includesDot
+        ? fullClassName.lastIndexOf(".") === idx && fullClassName.indexOf(":") === -1
+        : fullClassName.lastIndexOf(":") === idx)
+    );
+  }, `Invalid full class name: ${fullClassName}`);
+
+  return { schemaName: fullClassName.slice(0, idx), className: fullClassName.slice(idx + 1) };
 }
 
 /**